+++ modypy/blocks/rigid.py
@@ -2,7 +2,7 @@
 Blocks for stiff body dynamics
 """
 import numpy as np
-from modypy.model import Block, Port, SignalState, signal_method
+from modypy.model import Block, Port, signal_method, State
 from numpy import linalg
 
 
@@ -32,14 +32,16 @@ class RigidBody6DOFFlatEarth(Block):
         - the angular rates in the body reference frame.
     """
 
-    def __init__(self,
-                 owner,
-                 mass,
-                 moment_of_inertia,
-                 initial_velocity_earth=None,
-                 initial_position_earth=None,
-                 initial_transformation=None,
-                 initial_angular_rates_earth=None):
+    def __init__(
+        self,
+        owner,
+        mass,
+        moment_of_inertia,
+        initial_velocity_earth=None,
+        initial_position_earth=None,
+        initial_transformation=None,
+        initial_angular_rates_earth=None,
+    ):
         Block.__init__(self, owner)
         self.mass = mass
         self.moment_of_inertia = moment_of_inertia
@@ -50,55 +52,57 @@ class RigidBody6DOFFlatEarth(Block):
         if initial_transformation is None:
             initial_transformation = np.eye(3)
 
-        self.velocity_earth = \
-            SignalState(self,
-                        shape=3,
-                        derivative_function=self.velocity_earth_dot,
-                        initial_condition=initial_velocity_earth)
-        self.position_earth = \
-            SignalState(self,
-                        shape=3,
-                        derivative_function=self.position_earth_dot,
-                        initial_condition=initial_position_earth)
-        self.omega_earth = \
-            SignalState(self,
-                        shape=3,
-                        derivative_function=self.omega_earth_dot,
-                        initial_condition=initial_angular_rates_earth)
-        self.dcm = \
-            SignalState(self,
-                        shape=(3, 3),
-                        derivative_function=self.dcm_dot,
-                        initial_condition=initial_transformation)
+        self.velocity_earth = State(
+            self,
+            shape=3,
+            derivative_function=self.velocity_earth_dot,
+            initial_condition=initial_velocity_earth,
+        )
+        self.position_earth = State(
+            self,
+            shape=3,
+            derivative_function=self.position_earth_dot,
+            initial_condition=initial_position_earth,
+        )
+        self.omega_earth = State(
+            self,
+            shape=3,
+            derivative_function=self.omega_earth_dot,
+            initial_condition=initial_angular_rates_earth,
+        )
+        self.dcm = State(
+            self,
+            shape=(3, 3),
+            derivative_function=self.dcm_dot,
+            initial_condition=initial_transformation,
+        )
 
     def velocity_earth_dot(self, data):
-        """Calculates the acceleration in the earth reference frame
-        """
+        """Calculates the acceleration in the earth reference frame"""
         forces_earth = self.dcm(data) @ self.forces_body(data)
         accel_earth = forces_earth / self.mass
         return accel_earth
 
     def position_earth_dot(self, data):
-        """Calculates the velocity in the earth reference frame
-        """
+        """Calculates the velocity in the earth reference frame"""
         return self.velocity_earth(data)
 
     def omega_earth_dot(self, data):
-        """Calculate the angular acceleration in the earth reference frame
-        """
+        """Calculate the angular acceleration in the earth reference frame"""
         moments_earth = self.dcm(data) @ self.moments_body(data)
         ang_accel_earth = linalg.solve(self.moment_of_inertia, moments_earth)
         return ang_accel_earth
 
     def dcm_dot(self, data):
-        """Calculate the derivative of the direct cosine matrix
-        """
+        """Calculate the derivative of the direct cosine matrix"""
         omega_earth = self.omega_earth(data)
-        skew_sym_matrix = np.array([
-            [0, -omega_earth[2], omega_earth[1]],
-            [omega_earth[2], 0, -omega_earth[0]],
-            [-omega_earth[1], omega_earth[0], 0]
-        ])
+        skew_sym_matrix = np.array(
+            [
+                [0, -omega_earth[2], omega_earth[1]],
+                [omega_earth[2], 0, -omega_earth[0]],
+                [-omega_earth[1], omega_earth[0], 0],
+            ]
+        )
         return skew_sym_matrix @ self.dcm(data)
 
     # pylint does not recognize the modifications to the signal_method decorator
@@ -114,9 +118,11 @@ class RigidBody6DOFFlatEarth(Block):
     def velocity_body(self, data):
         """Calculate the velocity in the body reference frame"""
 
-        return np.einsum('ij...,j...->i...',
-                         self.dcm_inverse(data),
-                         self.velocity_earth(data))
+        return np.einsum(
+            "ij...,j...->i...",
+            self.dcm_inverse(data),
+            self.velocity_earth(data),
+        )
 
     # pylint does not recognize the modifications to the signal_method decorator
     # pylint: disable=no-value-for-parameter
@@ -124,9 +130,9 @@ class RigidBody6DOFFlatEarth(Block):
     def omega_body(self, data):
         """Calculate the angular velocity in the body reference frame"""
 
-        return np.einsum('ij...,j...->i...',
-                         self.dcm_inverse(data),
-                         self.omega_earth(data))
+        return np.einsum(
+            "ij...,j...->i...", self.dcm_inverse(data), self.omega_earth(data)
+        )
 
 
 class DirectCosineToEuler(Block):
@@ -147,21 +153,18 @@ class DirectCosineToEuler(Block):
 
     @signal_method
     def yaw(self, data):
-        """Calculate the yaw angle for the given direct cosine matrix
-        """
+        """Calculate the yaw angle for the given direct cosine matrix"""
         dcm = self.dcm(data)
         return np.arctan2(dcm[1, 0], dcm[0, 0])
 
     @signal_method
     def pitch(self, data):
-        """Calculate the pitch angle for the given direct cosine matrix
-        """
+        """Calculate the pitch angle for the given direct cosine matrix"""
         dcm = self.dcm(data)
         return np.arctan2(-dcm[2, 0], np.sqrt(dcm[0, 0] ** 2 + dcm[1, 0] ** 2))
 
     @signal_method
     def roll(self, data):
-        """Calculate the roll angle for the given direct cosine matrix
-        """
+        """Calculate the roll angle for the given direct cosine matrix"""
         dcm = self.dcm(data)
         return np.arctan2(dcm[2, 1], dcm[2, 2])